replacing the old master_main
